+++ Samples/TGPDemo/src/TGPListBoxEx.cpp
@@ -44,7 +44,7 @@ namespace DM
 			if (-1 == nIndex
 				||nIndex>(int)GetCount())
 			{
-				nIndex = GetCount();
+				nIndex = (int)GetCount();
 			}
 
 			LPTGPLBITEMEX pNewItem = new TGPLBITEMEX(XmlNode,this);
@@ -792,7 +792,7 @@ namespace DM
 			{
 				break;
 			}
-			int iItem = pPanel->GetItemId();
+			int iItem = (int)pPanel->GetItemId();
 			*lpRect = GetItemRect(iItem);
 			iErr = DM_ECODE_OK;
 		} while (false);
@@ -1048,7 +1048,7 @@ namespace DM
 					CRect rcItem(0,0,rcClient.Width(),m_DMArray[items[i]]->nHeight);
 					rcItem.OffsetRect(0,nTotalHeight-m_ptCurPos.y);
 					rcItem.OffsetRect(rcClient.TopLeft());
-					DrawItem(pCanvas,rcItem,items[i]);
+					DrawItem(pCanvas,rcItem,(int)items[i]);
 				}
 				nTotalHeight += m_DMArray[items[i]]->nHeight;
 			}
